+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Expose the result polling period and timeout as backend options #46
+* Support `qiskit.result.Result.get_memory()` to retrieve the raw results bitstrings #48
 
 ## qiskit-aqt-provider v0.10.0